+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## 0.1.0
 
 - Start Project:
-    - ...
+    - Capture rig setup.
+    - Example1.